+++ stencil.js
@@ -211,7 +211,7 @@ var stencil = {
             return newStencil;
         };
 
-        var tagElem = $("#" + tagID);
+        var tagElem = $("#" + stencil.util.escapeCSS(tagID));
 
         //Return a existing stencil if already initialized because template is already removed once initalized
         if(stencil.stencilHolder[tagID] != null) {
@@ -269,7 +269,7 @@ var stencil = {
         if(specificInners != null) {
             specificInners.forEach(function(innerTagID) {
                 stencil.util.log("Stencil: Building specified child stencil: " + innerTagID);
-                var tag = $("#" + innerTagID);
+                var tag = $("#" + stencil.util.escapeCSS(innerTagID));
                 if(!tag.length) {
                     stencil.util.log("Stencil: Specified child stencil: " + innerTagID + " not found!");
                 }
@@ -278,7 +278,7 @@ var stencil = {
                 //stencil, it will cause issues with ID selector as there are multiple tags with the same ID
                 //May consider removing innerTagID from ID to clean up
                 tag.addClass(innerTagID);
-                var innerDestination = "." + innerTagID; //Convert to JQuery class selector format
+                var innerDestination = "." + stencil.util.escapeCSS(innerTagID); //Convert to JQuery class selector format
                 
                 childStencils[innerTagID] = getStencil(innerTagID, innerDestination, tag.html());
                 childStencils[innerTagID].isStandard = false;
@@ -413,15 +413,71 @@ var stencil = {
                 unwrapper(this);
             });
         },
-        changeTag: function(elements, changeToType) {
-            //Don't use as it will replace all similar nodes within the HTML
-            elements.each(function(element) {
-                var replacementElement = this.outerHTML
-                                         .replace("/^<" + this.nodeName + "/", "<" + changeToType)
-                                         .replace("/<\/" + this.nodeName + ">$/", "</" + changeToType + ">");
-                element.replaceWith(replacementElement);
-            });
-            
+        escapeCSS: function(value) {
+            /*! http://mths.be/cssescape v0.2.1 by @mathias | MIT license */
+            // http://dev.w3.org/csswg/cssom/#serialize-an-identifier
+            var string = String(value);
+            var length = string.length;
+            var index = -1;
+            var codeUnit;
+            var result = '';
+            var firstCodeUnit = string.charCodeAt(0);
+            while (++index < length) {
+                codeUnit = string.charCodeAt(index);
+                // Note: there’s no need to special-case astral symbols, surrogate
+                // pairs, or lone surrogates.
+
+                // If the character is NULL (U+0000), then throw an
+                // `InvalidCharacterError` exception and terminate these steps.
+                if (codeUnit == 0x0000) {
+                    throw new InvalidCharacterError(
+                        'Invalid character: the input contains U+0000.'
+                    );
+                }
+
+                if (
+                    // If the character is in the range [\1-\1F] (U+0001 to U+001F) or is
+                    // U+007F, […]
+                    (codeUnit >= 0x0001 && codeUnit <= 0x001F) || codeUnit == 0x007F ||
+                    // If the character is the first character and is in the range [0-9]
+                    // (U+0030 to U+0039), […]
+                    (index == 0 && codeUnit >= 0x0030 && codeUnit <= 0x0039) ||
+                    // If the character is the second character and is in the range [0-9]
+                    // (U+0030 to U+0039) and the first character is a `-` (U+002D), […]
+                    (
+                        index == 1 &&
+                        codeUnit >= 0x0030 && codeUnit <= 0x0039 &&
+                        firstCodeUnit == 0x002D
+                    )
+                ) {
+                    // http://dev.w3.org/csswg/cssom/#escape-a-character-as-code-point
+                    result += '\\' + codeUnit.toString(16) + ' ';
+                    continue;
+                }
+
+                // If the character is not handled by one of the above rules and is
+                // greater than or equal to U+0080, is `-` (U+002D) or `_` (U+005F), or
+                // is in one of the ranges [0-9] (U+0030 to U+0039), [A-Z] (U+0041 to
+                // U+005A), or [a-z] (U+0061 to U+007A), […]
+                if (
+                    codeUnit >= 0x0080 ||
+                    codeUnit == 0x002D ||
+                    codeUnit == 0x005F ||
+                    codeUnit >= 0x0030 && codeUnit <= 0x0039 ||
+                    codeUnit >= 0x0041 && codeUnit <= 0x005A ||
+                    codeUnit >= 0x0061 && codeUnit <= 0x007A
+                ) {
+                    // the character itself
+                    result += string.charAt(index);
+                    continue;
+                }
+
+                // Otherwise, the escaped character.
+                // http://dev.w3.org/csswg/cssom/#escape-a-character
+                result += '\\' + string.charAt(index);
+
+            }
+            return result;
         },
         //Gets the value from a complex object via Javascript object notation
         getKeyValue: function(obj, keyString) {