+++ changelog.md
@@ -33,3 +33,7 @@ version 11
 1. Added string render output option to get pure string output
 2. Added solution for specificInners IDs which contain insertion objects e.g. {{lpIdx}}
 3. Updated solution for append render output option that doesn't refresh existing content that its appending to
+
+version 12
+-------------
+1. Updated depp object retrival notation to support array notation a.k.a. array[123].foo[bar]